% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cache.R
\name{cache_init}
\alias{cache_init}
\alias{cache_ls}
\alias{cache_dir}
\alias{cache_destroy}
\title{Data Caching API}
\usage{
cache_init(path = cache_dir())

cache_ls(path = cache_dir(), ...)

cache_dir()

cache_destroy(path = cache_dir(), verbose = TRUE)
}
\arguments{
\item{path}{character, path to cache directory, Default: cache_dir()}

\item{\dots}{arguments passed to \link[base:list.files]{list.files}}

\item{verbose}{logical, show message on success of \link{cache_destroy}}
}
\value{
character
}
\description{
Functions to manage cache data across sessions.
}
\examples{
 cache_dir()
 cache_init()
 cache_ls()
 cache_destroy()
}
