% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{cache_data}
\alias{cache_data}
\title{Store data}
\usage{
cache_data(data, filetype = "rds")
}
\arguments{
\item{data}{character, name of the object to store}

\item{filetype}{character, extension of the
filetype to save the data. Default: 'rds'}
}
\description{
Store data in the cache
}
\details{
Currently only rds is supported
}
\examples{
  # initialize the cache directory
  cache_init()

  # populate the cache
  cache_data('fev_data')
  cache_ls()

  # cleanup
   cache_destroy()
}
