% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cache.R
\name{cache_rm}
\alias{cache_rm}
\title{Clear data cache}
\usage{
cache_rm(files = cache_ls(full.names = TRUE), recursive = FALSE, force = FALSE)
}
\arguments{
\item{files}{character, files to be removed,
Default: \link{cache_ls}(full.names = TRUE)}

\item{recursive}{boolean. Should directories be deleted recursively?}

\item{force}{boolean. Should permissions be changed (if possible) to
allow the file or directory to be removed?}
}
\description{
Clear file(s) from the cache path.
}
\examples{
  # initialize the cache directory
  cache_init()

  # populate the cache
  cache_data("fev_data")
  cache_data("cars")

  # list files in the cache
  cache_ls()

  # clear all the files in the cache
  cache_rm()
  cache_ls()

 # remove a single file based on a pattern
  cache_data("fev_data")
  cache_data("cars")
  cache_rm(cache_ls(pattern = "fev", full.names = TRUE))
  cache_ls()

 # cleanup
  cache_destroy()
}
