% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_columns}
\alias{check_columns}
\title{Check Columns}
\usage{
check_columns(path)
}
\arguments{
\item{path}{\verb{directory path} Path to where \code{.csv} files are stored.}
}
\description{
Checks that columns map to the data specification in \code{{gsm}}.
}
\keyword{internal}
