% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_mapping}
\alias{check_mapping}
\title{Check Mapping}
\usage{
check_mapping(domain, data)
}
\arguments{
\item{domain}{\code{character} Data domain.}

\item{data}{\code{data.frame} Data.}
}
\description{
Checks that columns map to the data specification in \code{{gsm}}.
}
\keyword{internal}
