% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_rows}
\alias{check_rows}
\title{Check Rows}
\usage{
check_rows(df1, df2, domain = NULL)
}
\arguments{
\item{df1}{\code{data.frame} Data domain.}

\item{df2}{\code{data.frame} Data domain.}

\item{domain}{\code{character} Data domain.}
}
\description{
Check Rows
}
\keyword{internal}
