% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_protdev.R
\name{clean_protdev}
\alias{clean_protdev}
\title{Add \code{subjectid} to \code{protdev}}
\usage{
clean_protdev(protdev, dm)
}
\arguments{
\item{protdev}{\code{data.frame} protocol deviation data in rawplus format}

\item{dm}{\code{data.frame} subject-level data in rawplus format}
}
\description{
Add \code{subjectid} to \code{protdev}
}
