% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clindata-package.R
\docType{package}
\name{clindata-package}
\alias{clindata-package}
\alias{_PACKAGE}
\title{\code{clindata} Package}
\description{
\code{clindata} data package for clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/clindata}
  \item Report bugs at \url{https://github.com/openpharma/clindata/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Sidi \email{yoni@pinpointstrategies.io}

}
