% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{clindata_list}
\alias{clindata_list}
\title{FUNCTION_TITLE}
\usage{
clindata_list(filetype = c("csv", "json"))
}
\arguments{
\item{filetype}{PARAM_DESCRIPTION, Default: c("csv", "json")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
