% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config_schedule}
\alias{config_schedule}
\title{Snapshot Schedule}
\format{
a data frame with 75 rows and 2 columns
}
\source{
./data-raw/config/schedule.csv
}
\usage{
config_schedule
}
\description{
Snapshot Schedule (schedule) data CTMS data
}
\keyword{datasets}
