% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSUBJ.R
\name{CreateSUBJ}
\alias{CreateSUBJ}
\title{Utility Function to Create Raw Data Subject Level (RDSL)}
\usage{
CreateSUBJ(
  dfDm = NULL,
  dfIXRSrand = NULL,
  dfVisit = NULL,
  dfStud = NULL,
  dfEx = NULL,
  dfSdrg = NULL,
  dtSnapshot = NULL
)
}
\arguments{
\item{dfDm}{data frame of demographic information with required columns SUBJID INVID SCRNID}

\item{dfIXRSrand}{randomization dataset with required columns SUBJID RGMNDTC}

\item{dfVisit}{data frame of visit information with required columns SUBJID INVID FOLDERNAME RECORDDATE}

\item{dfStud}{optional, if no data frame is supplied will assume no subject completed study, otherwise
requires data frame of study completion information with columns SUBJID COMPYN_STD COMPREAS}

\item{dfEx}{data frame of treatment information with required columns SUBJID INVID EXSTDAT EXENDAT. If
multiple treatments and only want to focus on one treatment then input data frame will need to be subset
by user prior to input}

\item{dfSdrg}{optional, if no data frame is supplied will assume no subject completed treatment, otherwise
requires data frame of treatment completion information with columns SUBJID SDRGYN_STD. If multiple treatments
and only want to focus on one treatment then input data frame will need to be subset
by user prior to input}

\item{dtSnapshot}{date of data snapshot, if NULL, will impute to be current date}
}
\value{
dataframe with SubjectID, ScreenID, SiteID, ScreenFlag, RandFlag, RandDate, FirstStudyDate, LastStudyDate
TimeonStudy, FirstDoseDate, LastDoseDate, TimeOnTreatment
}
\description{
Create subject level data mapped from raw data for input
}
\details{
This output of this function is used by assessments as a standardized measurement of time on treatment.
}
\section{Data Specification}{


The following columns are required:
\itemize{
\item \code{dfDm} - Exposure raw dataset
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item \code{SCRNID} - Screening ID
}
\item \code{dfEX} - Exposure raw dataset
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item \code{EXSTDAT} - Start date of dose
\item \code{EXENDAT} - Stop date of dose
}
\item \code{dfIXRSrand} - IXRS randomization dataset
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{RGMNDTC} - randomization date
}
\item \code{dfVisit}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item \code{FOLDERNAME} - description of event, 'Screening' will not be considered for determining date.
\item \code{RECORDDATE} - Clinical Date of record (ex: visit date), Min and max per subject / site will be used to determine time on study (exposure)
}
}

The following columns are optional
\itemize{
\item \code{dfSdrg}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{SDRGYN_STD} - Y/N Did subject complete study drug closing
}
\item \code{dfStud}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{COMPYN_STD} - Y/N Did subject complete duration of study
\item \code{COMPREAS} - character, if COMPYN_STD = N, reason for study discontinue.
}

-\code{dfSnapshot} - Date, Date of snapshot
}
}

\examples{

rawplus_SUBJ <- CreateSUBJ(
  dfDm = clindata::raw_dm,
  dfIXRSrand = clindata::raw_iwrsrand,
  dfEx = clindata::raw_ex,
  dfVisit = clindata::raw_visdt,
  dfStud = clindata::raw_studcomp,
  dfSdrg = clindata::raw_sdrgcomp
)

}
