% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ctms.R
\name{ctms_1_import}
\alias{ctms_1_import}
\title{Import CTMS data}
\usage{
ctms_1_import(data_path = file.path("data-raw", "ctms"))
}
\arguments{
\item{data_path}{\code{character} path to data files}
}
\description{
Import CTMS data
}
\keyword{internal}
