% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms_make.R
\name{ctms_3_export}
\alias{ctms_3_export}
\title{Export CTMS data}
\usage{
ctms_3_export(datasets_processed, data_domain = "ctms")
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}
}
\description{
Export CTMS data
}
\keyword{internal}
