% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms_4_document.R
\name{ctms_4_document}
\alias{ctms_4_document}
\title{Document CTMS data}
\usage{
ctms_4_document(
  datasets_processed,
  domain_metadata_path = system.file("data-standards", "ctms.yaml", package =
    "clindata"),
  documentation_path = paste0(system.file("R", package = "clindata"), "/ctms.R"),
  data_domain = "ctms"
)
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}

\item{domain_metadata_path}{\code{character} path to domain metadata file}

\item{documentation_path}{\code{character} path to documentation file}
}
\description{
Document CTMS data
}
