% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms.R
\docType{data}
\name{ctms_protdev}
\alias{ctms_protdev}
\title{Protocol Deviations}
\format{
a data frame with 4646 rows and 4 columns
}
\source{
./data-raw/ctms/protdev.csv
}
\usage{
ctms_protdev
}
\description{
Protocol Deviations (protdev) data CTMS data
}
\keyword{datasets}
