% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms.R
\docType{data}
\name{ctms_site}
\alias{ctms_site}
\title{Site Metadata}
\format{
a data frame with 176 rows and 14 columns
}
\source{
./data-raw/ctms/site.csv
}
\usage{
ctms_site
}
\description{
Site Metadata (site) data CTMS data
}
\keyword{datasets}
