% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms.R
\docType{data}
\name{ctms_study}
\alias{ctms_study}
\title{Study Metadata}
\format{
a data frame with 1 rows and 10 columns.
}
\usage{
ctms_study
}
\description{
Study Metadata (study) CTMS data.
}
\keyword{datasets}
