% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(data, filetype = "rds", use_cache = FALSE)
}
\arguments{
\item{data}{character, name of data object}

\item{filetype}{character, extension type of the data. Default: "rds"}

\item{use_cache}{boolean, use the data stored in cache? Default: FALSE}
}
\value{
data stored in either the local cache directory or within the package.
}
\description{
Read data directly into a global environment
}
\details{
Currently only loading of rds files is supported
}
\examples{
 read_data("fev_data")
 read_data("bcva_data")
}
