% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc_1_import.R
\name{edc_1_import}
\alias{edc_1_import}
\title{Import EDC data}
\usage{
edc_1_import(data_path = file.path("data-raw", "edc"))
}
\arguments{
\item{data_path}{\code{character} path to data files}
}
\description{
Import EDC data
}
