% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc_3_export.R
\name{edc_3_export}
\alias{edc_3_export}
\title{Export EDC data}
\usage{
edc_3_export(datasets_processed)
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}
}
\description{
Export EDC data
}
