% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc_make.R
\name{edc_4_document}
\alias{edc_4_document}
\title{Document EDC data}
\usage{
edc_4_document(
  datasets_processed,
  domain_metadata_path = system.file("data-standards", "edc.yaml", package = "clindata"),
  documentation_path = paste0(system.file("R", package = "clindata"), "/edc.R")
)
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}

\item{domain_metadata_path}{\code{character} path to domain metadata file}

\item{documentation_path}{\code{character} path to documentation file}
}
\description{
Document EDC data
}
\keyword{internal}
