% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc.R
\docType{data}
\name{edc_data_change_rate}
\alias{edc_data_change_rate}
\title{Data Change Rate}
\format{
a data frame with 240398 rows and 7 columns
}
\source{
./data-raw/edc/data_change_rate.csv
}
\usage{
edc_data_change_rate
}
\description{
data page-level data with simulated data changes (data_change_rate)
}
\keyword{datasets}
