% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc.R
\docType{data}
\name{edc_data_pages}
\alias{edc_data_pages}
\title{Data Entry Lag}
\format{
a data frame with 247462 rows and 6 columns
}
\source{
./data-raw/edc/data_pages.csv
}
\usage{
edc_data_pages
}
\description{
data page-level data with simulated data entry lag (data_pages)
}
\keyword{datasets}
