% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc.R
\docType{data}
\name{edc_data_points}
\alias{edc_data_points}
\title{Data Points}
\format{
a data frame with 2114862 rows and 10 columns
}
\source{
./data-raw/edc/data_points.csv
}
\usage{
edc_data_points
}
\description{
data point-level data with simulated count of data changes (data_points)
}
\keyword{datasets}
