% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-FlagTELab.R
\name{FlagTELab}
\alias{FlagTELab}
\title{Utility Function to Identify Treatment Emergent Lab Abnormalities}
\usage{
FlagTELab(dfLab, dfRDSL, strToxVar = "TOXGRG")
}
\arguments{
\item{dfLab}{data frame of raw lab information}

\item{dfRDSL}{data frame of rawplus subject level derived info}

\item{strToxVar}{variable name of toxicity grade in dfLab dataset, default is TOXGRG}
}
\value{
original lab dataset with additional columns for BLTOX (baseline toxicity grade) and TOXFLG (lab abnorm flag)
}
\description{
Creates two columns for the baseline toxicity grade and flagging for treatment emergent lab abnormalities
}
\details{
baseline toxicity grade is defined as the closest lab measurement to participant's first dose date
treatment emergent lab abnormalities are defined as graded abnormalities (tox grade > 0) after the first dose date
and wrose than the baseline toxicity grade
}
\section{Data Specification}{


The following datasets and column are required:
\itemize{
\item \code{dfLab}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item TO be completed
}
\item \code{dfRDSL}
\itemize{
\item \code{SUBJID} - Unique subject ID
\item \code{INVID} - Unique Investigator ID
\item To be completed
}
}
}

\examples{

FlagTELab(dfLab = clindata::raw_covlab_hema, dfRDSL = clindata::rawplus_rdsl)

}
