% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{get_snapshot_date}
\alias{get_snapshot_date}
\title{Returns an earlier date}
\usage{
get_snapshot_date(
  snapshot_date = NULL,
  interval = lubridate::month,
  n_intervals = 1
)
}
\arguments{
\item{snapshot_date}{\code{Date} Initial snapshot date}

\item{interval}{\code{function} Interval size as implemented in {lubridate}, one of \code{year}, \code{month}, or \code{day}}

\item{n_intervals}{\code{numeric} Number of intervals}
}
\value{
\code{Date}
}
\description{
Returns an earlier date
}
