% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(data, env = parent.frame(), ...)
}
\arguments{
\item{data}{character, name of data object}

\item{env}{environment, Default: parent.frame()}

\item{...}{arguments passed to assign \link[base:assign]{assign}}
}
\description{
Load data directly into an enivornment
}
\details{
By default the data is loaded into the parent frame from which it is called.
}
\examples{
 load_data("fev_data")
}
