% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRawplus.R
\name{map_rawplus_ae}
\alias{map_rawplus_ae}
\title{map_rawplus_ae}
\usage{
map_rawplus_ae(dfRawAE, ids)
}
\arguments{
\item{dfRawAE}{data frame of raw AE data}
}
\description{
mapping to create rawplus_ae from raw
}
\examples{

rawplus_ae <- map_rawplus_ae(clindata::raw_ae)

}
