% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRawplus.R
\name{map_rawplus_consent}
\alias{map_rawplus_consent}
\title{map_rawplus_consent}
\usage{
map_rawplus_consent(dfRawElig, ids)
}
\arguments{
\item{dfRawElig}{data frame of raw AE data}
}
\description{
mapping to create map_rawplus_consent from raw
}
\examples{

rawplus_consent <- map_rawplus_consent(clindata::raw_ic_elig)
}
