% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRawplus.R
\name{map_rawplus_ie}
\alias{map_rawplus_ie}
\title{map_rawplus_ie}
\usage{
map_rawplus_ie(dfRawIE, ids)
}
\arguments{
\item{dfRawIE}{data frame of raw AE data}
}
\description{
mapping to create rawplus_ie from raw
}
\examples{

rawplus_ie <- map_rawplus_ie(clindata::raw_ie_all)

}
