% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\docType{data}
\name{raw_covlab_hema}
\alias{raw_covlab_hema}
\title{raw_covlab_hema}
\format{
a data frame with 531142 rows and 72 columns.
}
\usage{
raw_covlab_hema
}
\description{
Raw covlab_hema data frame
}
\keyword{datasets}
