% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\docType{data}
\name{raw_wdconsent}
\alias{raw_wdconsent}
\title{raw_wdconsent}
\format{
a data frame with 29 rows and 26 columns.
}
\usage{
raw_wdconsent
}
\description{
Raw wdconsent data frame
}
\keyword{datasets}
