% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus_1_import.R
\name{rawplus_1_import}
\alias{rawplus_1_import}
\title{Import \link{ rawplus } data}
\usage{
rawplus_1_import(data_path = file.path("data-raw", "rawplus"))
}
\arguments{
\item{data_path}{\code{character} path to data files}
}
\description{
Import \link{ rawplus } data
}
