% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rawplus.R
\name{rawplus_2_process_1_dm}
\alias{rawplus_2_process_1_dm}
\title{Process \link{ dm } domain}
\usage{
rawplus_2_process_1_dm(dm)
}
\arguments{
\item{dm}{\code{data.frame} \link{ dm } data}
}
\description{
Process \link{ dm } domain
}
\keyword{internal}
