% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus_make.R
\name{rawplus_2_process_2_domain}
\alias{rawplus_2_process_2_domain}
\title{Process domain data}
\usage{
rawplus_2_process_2_domain(data, dm, domain)
}
\arguments{
\item{data}{\code{data.frame} domain data}

\item{dm}{\code{data.frame} \link{ dm } data}

\item{domain}{\code{character} domain name}
}
\description{
Remove rows from domain data where the associated subject ID does not exist in \link{ dm }.
}
\keyword{internal}
