% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus_2_process_3_lb.R
\name{rawplus_2_process_3_lb}
\alias{rawplus_2_process_3_lb}
\title{Process \link{ lb } data}
\usage{
rawplus_2_process_3_lb(lb, dm)
}
\arguments{
\item{lb}{\code{data.frame} \link{ lb } data}

\item{dm}{\code{data.frame} \link{ dm } data}
}
\description{
Reduce the size of \link{ lb } data.
}
