% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rawplus.R
\name{rawplus_3_export}
\alias{rawplus_3_export}
\title{Export \link{ rawplus } data}
\usage{
rawplus_3_export(datasets_processed)
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}
}
\description{
Export \link{ rawplus } data
}
\keyword{internal}
