% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus_4_document.R
\name{rawplus_4_document}
\alias{rawplus_4_document}
\title{Document \link{ rawplus } data}
\usage{
rawplus_4_document(
  datasets_processed,
  domain_metadata_path = system.file("data-standards", "rawplus.yaml", package =
    "clindata"),
  column_metadata_path = file.path("data-raw", "rawplus"),
  documentation_path = paste0(system.file("R", package = "clindata"), "/rawplus.R")
)
}
\arguments{
\item{datasets_processed}{\code{list} named list of processed data frames}

\item{domain_metadata_path}{\code{character} path to domain metadata file}

\item{column_metadata_path}{\code{character} path to column metadata files}

\item{documentation_path}{\code{character} path to documentation file}
}
\description{
Document \link{ rawplus } data
}
