% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_consent}
\alias{rawplus_consent}
\title{rawplus_consent}
\format{
a data frame with 1297 rows and 4 columns.
}
\source{
./data-raw/rawplus/rawplus_consent.rda
}
\usage{
rawplus_consent
}
\description{
Visits (visdt) data in \code{rawplus} format.
\describe{
\item{subjid}{\code{string} Subject Identifier}
\item{consdt}{\code{date} Date of Consent}
\item{conscat}{\code{string} Individual giving consent}
\item{consyn}{\code{character} Consent Indication}
}
}
\keyword{datasets}
