% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_enroll}
\alias{rawplus_enroll}
\title{rawplus_enroll}
\format{
a data frame with 2387 rows and 9 columns.
}
\source{
./data/rawplus_enroll.rda
}
\usage{
rawplus_enroll
}
\description{
Visits (visdt) data in \code{rawplus} format.
\describe{
\item{studyid}{\code{string} Study Identifier}
\item{siteid}{\code{string} Site Identifier}
\item{subjectid}{\code{string} Detailed subject Identifier}
\item{subjid}{\code{string} Trial Subject Identifier}
\item{enroll_dt}{\code{date} Date of enrollment in clinical trial}
\item{enrollyn}{\code{character} Indicator of enrollment in clinical trial, Y = yes, N = no}
\item{sfreas}{\code{string} Reason for withrawl from trial}
\item{country}{\code{string} Country of trial}
\item{invid}{\code{string} Subject Identifier}
}
}
\keyword{datasets}
