% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_ex}
\alias{rawplus_ex}
\title{rawplus_ex}
\format{
a data frame with 4876 rows and 53 columns.
}
\source{
./data-raw/rawplus/ex.parquet
}
\usage{
rawplus_ex
}
\description{
Exposure (ex) data in \code{rawplus} format.
\describe{
\item{studyid}{\code{string} Study Identifier}
\item{siteid}{\code{string} Site Identifier}
\item{invid}{\code{string} Investigator ID}
\item{scrnid}{\code{string} Screen Identifier}
\item{subjid}{\code{string} Subject Identifier}
\item{subjectid}{\code{string} System Subject Identifier}
\item{datapagename}{\code{string} eCRF Page Name}
\item{datapageid}{\code{string} Internal ID for Data Page}
\item{foldername}{\code{string} Folder Name}
\item{instancename}{\code{string} Instance Name}
\item{recordid}{\code{string} Record Identifier}
\item{record_dt}{\code{string} Record Date}
\item{recordposition}{\code{float} Record Position}
\item{mincreated_dts}{\code{string} Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\code{string} Latest Data Update Date/Time}
\item{extrt}{\code{string} Treatment Name}
\item{exds}{\code{string} Actual Dose}
\item{exdsu}{\code{string} Acutal Dose Units}
\item{exdsfrmn}{\code{int} Number of Dose Form}
\item{exdsfrm}{\code{string} Dose Form}
\item{exdsfrq}{\code{string} Dosing Frequecy}
\item{exdsfrot}{\code{string} Other Dosing Frequecy}
\item{exrte}{\code{string} Route}
\item{exst_dt}{\code{string} Start Date/Time of Treatment}
\item{exen_dt}{\code{string} End Date/Time of Treatment}
\item{exongo}{\code{string} EX Ongoing}
\item{experm}{\code{string} Permanently Withdrawn}
\item{subjid_nsv}{\code{string} SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\code{string} SCREENING IDENTIFIER}
\item{subjinit_nsv}{\code{string} SUBJECT INITIALS}
\item{invid_nsv}{\code{string} INVESTIGATOR ID}
\item{subject_nsv}{\code{string} Subject name or identifier}
\item{instanceid_nsv}{\code{double} Internal id for the instance}
\item{folder_nsv}{\code{string} Folder OID}
\item{folderseq_nsv}{\code{double} Folder sequence number}
\item{extrt_std_nsv}{\code{string} Drug Name_Coded Value}
\item{exdstxt_nsv}{\code{double} Dose per administration}
\item{exdosu_std_nsv}{\code{double} Dose Units_Coded Value}
\item{exdsfrmn_raw_nsv}{\code{string} Number of insert Dose Form (Tablets)_(Character)}
\item{exdosfrm_std_nsv}{\code{double} Dose Form_Coded Value}
\item{exdosfrq_std_nsv}{\code{string} Dosing Frequency_Coded Value}
\item{exroute_std_nsv}{\code{double} Route_Coded Value}
\item{exrtoth_nsv}{\code{string} EXRTOTH}
\item{exstdat_nsv}{\code{timestamp} Start Date/Dose Date}
\item{exstdat_yy_nsv}{\code{double} Start Date/Dose Date_Year}
\item{exstdat_mm_nsv}{\code{double} Start Date/Dose Date_Month}
\item{exstdat_dd_nsv}{\code{double} Start Date/Dose Date_Day}
\item{exendat_nsv}{\code{timestamp} Stop Date}
\item{exendat_yy_nsv}{\code{double} Stop Date_Year}
\item{exendat_mm_nsv}{\code{double} Stop Date_Month}
\item{exendat_dd_nsv}{\code{double} Stop Date_Day}
\item{exongo_std_nsv}{\code{string} Check if Ongoing_Coded Value}
\item{experm_nsv}{\code{string} if study drug was permanently withdrawn}
}
}
\keyword{datasets}
