% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_ie}
\alias{rawplus_ie}
\title{rawplus_ie}
\format{
a data frame with 30168 rows and 4 columns.
}
\usage{
rawplus_ie
}
\description{
Inclusion/exclusion data in Raw+ format
}
\keyword{datasets}
