% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_ixrsrand}
\alias{rawplus_ixrsrand}
\title{rawplus_ixrsrand}
\format{
a data frame with 1301 rows and 18 columns.
}
\source{
./data-raw/rawplus/ixrsrand.parquet
}
\usage{
rawplus_ixrsrand
}
\description{
Randomization (ixrsrand) data in \code{rawplus} format.
\describe{
\item{studyid}{\code{string}}
\item{invid}{\code{string}}
\item{invnam}{\code{string}}
\item{oinvid}{\code{string}}
\item{oinvnam}{\code{string}}
\item{country}{\code{string}}
\item{ocountry}{\code{string}}
\item{txtyp_nsv}{\code{string}}
\item{scrnid_nsv}{\code{string}}
\item{subjid}{\code{string}}
\item{subjinit_nsv}{\code{string}}
\item{hbvdna_nsv}{\code{string}}
\item{virtreat_nsv}{\code{string}}
\item{trtgrpc}{\code{string}}
\item{randnum}{\code{string}}
\item{dispid_nsv}{\code{double}}
\item{status}{\code{string}}
\item{rand_dt}{\code{string}}
}
}
\keyword{datasets}
