% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_lb}
\alias{rawplus_lb}
\title{rawplus_lb}
\format{
a data frame with 1132070 rows and 14 columns.
\describe{
\item{subjid}{\link{ string } Subject ID or Number}
\item{visnam}{\link{ string } Visit Name}
\item{visnum}{\link{ float } Visit ID or Number}
\item{lb_dt}{\link{ string } Actual Collection Date and Time}
\item{battrnam}{\link{ string } Battery Name}
\item{lbtstnam}{\link{ string } Test Name ID}
\item{lbtstcd}{\link{ string } Test ID}
\item{siresn}{\link{ float } SI Numeric Result}
\item{sinrlo}{\link{ string } SI Reference Range Low}
\item{sinrhi}{\link{ string } SI Reference Range High}
\item{toxgr}{\link{ string } Toxicity Grade}
\item{alrtfl}{\link{ string } Alert Flag}
\item{treatmentemergent}{\link{ character }}
\item{alertsimplified}{\link{ character }}
}
}
\source{
./data-raw/rawplus/lb.parquet
}
\usage{
rawplus_lb
}
\description{
Labs (lb) data in \link{ rawplus } format.
}
\keyword{datasets}
