% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_pd}
\alias{rawplus_pd}
\title{rawplus_pd}
\format{
a data frame with 4459 rows and 4 columns.
}
\usage{
rawplus_pd
}
\description{
Protocol deviation data in Raw+ format
}
\keyword{datasets}
