% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_protdev}
\alias{rawplus_protdev}
\title{rawplus_protdev}
\format{
a data frame with 4837 rows and 26 columns.
\describe{
\item{invid}{\link{ string } Investigator ID}
\item{scrnid}{\link{ string } Screen Identifier}
\item{subjid}{\link{ string } Subject Identifier}
\item{dv_dt}{\link{ string } Date of Deviation}
\item{dvdecod}{\link{ string } PD Category}
\item{dvterm}{\link{ string } PD Description}
\item{importnt}{\link{ string } Deemed Important}
\item{cro_nsv}{\link{ string } CRO}
\item{country_nsv}{\link{ string } Country}
\item{invnam_nsv}{\link{ string } Principal Investigator}
\item{subjinit_nsv}{\link{ string } Initials}
\item{idendate_nsv}{\link{ string } Date Deviation Identified}
\item{visit_nsv}{\link{ string } Study Visit}
\item{devdesc_nsv}{\link{ string } Description}
\item{devused_nsv}{\link{ string } Protocol Deviation Form (GF-23005A) Used?}
\item{repdatec_nsv}{\link{ string } Date of Report on PD Form}
\item{comments_nsv}{\link{ string } Comments}
\item{gsidatec_nsv}{\link{ string } Date Sent to GSI}
\item{mmdatec_nsv}{\link{ string } Date Signed by  GSI MM}
\item{addatec_nsv}{\link{ string } Date Signed by GSI AD}
\item{retdatec_nsv}{\link{ string } Date Returned to CRO}
\item{tmfdatec_nsv}{\link{ string } Date Completed Form filed in TMF}
\item{mmpddatec_nsv}{\link{ string } Gilead MM Review Date of PD}
\item{revwdatec_nsv}{\link{ string } Date  of Review Date}
\item{devdatec_nsv}{\link{ string } Date of Deviation}
\item{idendatec_nsv}{\link{ string } Date Deviation Identified}
}
}
\source{
C:/Users/schildress/dev/clindata/data-raw/rawplus/protdev.parquet
}
\usage{
rawplus_protdev
}
\description{
Protocol Deviations (protdev) data in Raw+ format.
}
\keyword{datasets}
