% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_rdsl}
\alias{rawplus_rdsl}
\title{rawplus_rdsl}
\format{
a data frame with 2387 rows and 16 columns.
}
\usage{
rawplus_rdsl
}
\description{
raw data subject level (rdsl) data using Raw Plus format
}
\keyword{datasets}
