% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_sdrgcomp}
\alias{rawplus_sdrgcomp}
\title{rawplus_sdrgcomp}
\format{
a data frame with 689 rows and 35 columns.
}
\source{
./data-raw/rawplus/sdrgcomp.parquet
}
\usage{
rawplus_sdrgcomp
}
\description{
Treatment Disposition (sdrgcomp) data in \code{rawplus} format.
\describe{
\item{studyid}{\code{string} Study Identifier}
\item{siteid}{\code{string} Site Identifier}
\item{invid}{\code{string} Investigator ID}
\item{scrnid}{\code{string} Screen Identifier}
\item{subjid}{\code{string} Subject Identifier}
\item{subjectid}{\code{string} System Subject Identifier}
\item{datapagename}{\code{string} eCRF Page Name}
\item{datapageid}{\code{string} Internal ID for Data Page}
\item{foldername}{\code{string} Folder Name}
\item{instancename}{\code{string} Instance Name}
\item{recordid}{\code{string} Record Identifier}
\item{record_dt}{\code{string} Record Date}
\item{recordposition}{\code{float} Record Position}
\item{mincreated_dts}{\code{string} Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\code{string} Latest Data Update Date/Time}
\item{sdrgyn}{\code{string} Did subject complete study drug dosing?}
\item{sdrgreas}{\code{string} Reason for study drug discontinue}
\item{sdrgterm}{\code{string} Discontinuation reasons}
\item{subjid_nsv}{\code{string} SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\code{string} SCREENING IDENTIFIER}
\item{subjinit_nsv}{\code{string} SUBJECT INITIALS}
\item{invid_nsv}{\code{string} INVESTIGATOR ID}
\item{subject_nsv}{\code{string} Subject name or identifier}
\item{instanceid_nsv}{\code{double} Internal id for the instance}
\item{folder_nsv}{\code{string} Folder OID}
\item{folderseq_nsv}{\code{double} Folder sequence number}
\item{sdrgyn_std_nsv}{\code{string} Open label study drug completed_Coded Value}
\item{sdrgreas_std_nsv}{\code{string} Reason for study drug discontinuation_Coded Value}
\item{sdrgol_nsv}{\code{string} Will subject go to Open-Label}
\item{sdrgol_std_nsv}{\code{string} Will subject go to Open-Label_Coded Value}
\item{sdrgfu_nsv}{\code{string} Will subject enter 24-week trt free fup}
\item{sdrgfu_std_nsv}{\code{string} Will subject enter 24-week trt free fup_Coded Value}
\item{sdrgtrt_nsv}{\code{string} Did subject start another HBV therapy}
\item{sdrgtrt_std_nsv}{\code{string} Did subject start another HBV therapy_Coded Value}
\item{phase}{\code{character}}
}
}
\keyword{datasets}
