% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_studcomp}
\alias{rawplus_studcomp}
\title{rawplus_studcomp}
\format{
a data frame with 133 rows and 31 columns.
}
\source{
./data-raw/rawplus/studcomp.parquet
}
\usage{
rawplus_studcomp
}
\description{
Study Disposition (studcomp) data in \code{rawplus} format.
\describe{
\item{studyid}{\code{string} Study Identifier}
\item{siteid}{\code{string} Site Identifier}
\item{invid}{\code{string} Investigator ID}
\item{scrnid}{\code{string} Screen Identifier}
\item{subjid}{\code{string} Subject Identifier}
\item{subjectid}{\code{string} System Subject Identifier}
\item{datapagename}{\code{string} eCRF Page Name}
\item{datapageid}{\code{string} Internal ID for Data Page}
\item{foldername}{\code{string} Folder Name}
\item{instancename}{\code{string} Instance Name}
\item{recordid}{\code{string} Record Identifier}
\item{record_dt}{\code{string} Record Date}
\item{recordposition}{\code{float} Record Position}
\item{mincreated_dts}{\code{string} Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\code{string} Latest Data Update Date/Time}
\item{compyn}{\code{string} Subject complete duration of the study?}
\item{compreas}{\code{string} Reason for study discontinue}
\item{subjid_nsv}{\code{string} SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\code{string} SCREENING IDENTIFIER}
\item{subjinit_nsv}{\code{string} SUBJECT INITIALS}
\item{invid_nsv}{\code{string} INVESTIGATOR ID}
\item{subject_nsv}{\code{string} Subject name or identifier}
\item{instanceid_nsv}{\code{double} Internal id for the instance}
\item{folder_nsv}{\code{string} Folder OID}
\item{folderseq_nsv}{\code{double} Folder sequence number}
\item{compyn_std_nsv}{\code{string} Subject complete duration of the study?_Coded Value}
\item{compreas_std_nsv}{\code{string} If No, reason for study discontinuation_Coded Value}
\item{compfu_nsv}{\code{string} 24-week trt free fup completed}
\item{compfu_std_nsv}{\code{string} 24-week trt free fup completed_Coded Value}
\item{comptrt_nsv}{\code{string} Did subject start another HBV therapy}
\item{comptrt_std_nsv}{\code{string} Did subject start another HBV therapy_Coded Value}
}
}
\keyword{datasets}
