% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_subj}
\alias{rawplus_subj}
\title{rawplus_subj}
\format{
a data frame with 1297 rows and 18 columns.
}
\usage{
rawplus_subj
}
\description{
raw data subject level (rdsl) data using Raw Plus format
}
\keyword{datasets}
