% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{clindata_fetch}
\alias{clindata_fetch}
\alias{clindata_load}
\alias{clindata_read}
\title{FUNCTION_TITLE}
\usage{
clindata_fetch(file_name, filetype = c("csv", "json"), ...)

clindata_load(key, env = parent.frame(), ...)

clindata_read(key, filetype = c("csv", "json"))
}
\arguments{
\item{file_name}{PARAM_DESCRIPTION}

\item{filetype}{PARAM_DESCRIPTION, Default: c("csv", "json")}

\item{\dots}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[jsonlite:read_json]{read_json}
\link[readr:read_delim]{read_csv}
}
