% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_points.R
\name{simulate_data_points}
\alias{simulate_data_points}
\title{TODO: simulate data points (currently just code from simulate_data_pages)}
\usage{
simulate_data_points(
  dm,
  data_points = clindata::edc_data_points,
  data_page_rate = runif(1, 0.1, 0.4)
)
}
\description{
TODO: simulate data points (currently just code from simulate_data_pages)
}
