% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_study.R
\name{simulate_study}
\alias{simulate_study}
\title{Simulate study data}
\usage{
simulate_study(
  n_sites = NULL,
  n_subjects = NULL,
  start_date = NULL,
  end_date = NULL,
  rename_gsm = FALSE
)
}
\arguments{
\item{n_sites}{\code{integer} Number of sites}

\item{n_subjects}{\code{integer} Number of subjects to be divided among sites}

\item{start_date}{\code{Date} Start date of study}

\item{end_date}{\code{Date} End date of study}

\item{rename_gsm}{\code{logical} Apply {gsm} domain names?}
}
\value{
\code{list} Named list with simulated study data.
}
\description{
Simulate study data based on desired number of sites, number of subjects, and study duration.
}
