% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_ae}
\alias{snapshot_ae}
\title{Snapshot AE domain.}
\usage{
snapshot_ae(
  snapshot_date,
  dm,
  ae = clindata::rawplus_ae,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data.}

\item{dm}{\code{data.frame} Demographic data.}

\item{ae}{\code{data.frame} Adverse event data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot AE domain.
}
