% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_all.R
\name{snapshot_all}
\alias{snapshot_all}
\title{Snapshot all data domains}
\usage{
snapshot_all(
  snapshot_date = get_snapshot_date(),
  data = list(visdt = clindata::rawplus_visdt, ex = clindata::rawplus_ex, dm =
    clindata::rawplus_dm, studcomp = clindata::rawplus_studcomp, sdrgcomp =
    clindata::rawplus_sdrgcomp, lb = clindata::rawplus_lb, ae = clindata::rawplus_ae,
    protdev = clindata::ctms_protdev, enroll = clindata::rawplus_enroll, queries =
    clindata::edc_queries, data_pages = clindata::edc_data_pages, data_points =
    clindata::edc_data_points),
  impute_rf_dt = TRUE,
  print_check_rows = TRUE,
  rename_gsm = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{data}{\code{list} Name list of data domains to snapshot}

\item{impute_rf_dt}{\code{logical} Imput reference dates with \link{ visdt } and \link{ ex }?}
}
\value{
\code{list} Named list of modified data domains
}
\description{
Snapshot all data domains at a specific point in time.
}
