% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_consent}
\alias{snapshot_consent}
\title{Snapshot Consent Domain}
\usage{
snapshot_consent(
  snapshot_date,
  dm,
  consent = clindata::rawplus_consent,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data. \code{data.frame} Demographic data.}

\item{consent}{\code{data.frame} Consent data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}. \code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Consent Domain
}
