% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_data_pages}
\alias{snapshot_data_pages}
\title{Snapshot Data Pages Domain}
\usage{
snapshot_data_pages(
  snapshot_date,
  data_pages = clindata::edc_data_pages,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{data_pages}{\code{data.frame} EDC data page data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Data Pages Domain
}
