% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_data_points}
\alias{snapshot_data_points}
\title{Snapshot Data Points Domain}
\usage{
snapshot_data_points(
  snapshot_date,
  data_points = clindata::edc_data_points,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{data_points}{\code{data.frame} EDC Data Points data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Data Points Domain
}
