% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_dm}
\alias{snapshot_dm}
\title{Snapshot Demographic Domain}
\usage{
snapshot_dm(
  snapshot_date,
  visdt = NULL,
  ex = NULL,
  dm = clindata::rawplus_dm,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{visdt}{\code{Date} Visit date.}

\item{ex}{\code{data.frame} #TODO: complete documentation.}

\item{dm}{\code{data.frame} Demographic data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Demographic Domain
}
