% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_enroll}
\alias{snapshot_enroll}
\title{Snapshot Enroll Domain}
\usage{
snapshot_enroll(
  snapshot_date,
  enroll = clindata::rawplus_enroll,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{enroll}{\code{data.frame} Enrollment data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Enroll Domain
}
