% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_ex}
\alias{snapshot_ex}
\title{Snapshot EX Domain}
\usage{
snapshot_ex(snapshot_date, ex = clindata::rawplus_ex, print_check_rows = TRUE)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{ex}{\code{data.frame} data}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot EX Domain
}
