% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_ie}
\alias{snapshot_ie}
\title{Snapshot Inclusion/Exclusion Domain.}
\usage{
snapshot_ie(
  snapshot_date,
  dm,
  ie = clindata::rawplus_ie,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{ie}{\code{data.frame} Inclusion/exclusion data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Inclusion/Exclusion Domain.
}
