% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_lb}
\alias{snapshot_lb}
\title{Snapshot Labs Domain}
\usage{
snapshot_lb(
  snapshot_date,
  dm,
  lb = clindata::rawplus_lb,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{lb}{\code{data.frame} Lab values data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Labs Domain
}
