% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_protdev}
\alias{snapshot_protdev}
\title{Snapshot Protocol Deviation Domain}
\usage{
snapshot_protdev(
  snapshot_date,
  dm,
  protdev = clindata::ctms_protdev,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{protdev}{\code{data.frame} Protocol deviation data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Protocol Deviation Domain
}
