% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_queries}
\alias{snapshot_queries}
\title{Snapshot Queries Domain}
\usage{
snapshot_queries(
  snapshot_date,
  queries = clindata::edc_queries,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{queries}{\code{data.frame} EDC queries data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot Queries Domain
}
