% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_sdrgcomp}
\alias{snapshot_sdrgcomp}
\title{Snapshot SDRCOMP Domain}
\usage{
snapshot_sdrgcomp(
  snapshot_date,
  dm,
  sdrgcomp = clindata::rawplus_sdrgcomp,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{sdrgcomp}{\code{data.frame} SDRGCOMP data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot SDRCOMP Domain
}
