% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_site}
\alias{snapshot_site}
\title{Snapshot CTMS Site Domain}
\usage{
snapshot_site(
  snapshot_date,
  dm = clindata::rawplus_dm,
  site = clindata::ctms_site,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{site}{\code{data.frame} CTMS site data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot CTMS Site Domain
}
