% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_studcomp}
\alias{snapshot_studcomp}
\title{Snapshot STUDCOMP Domain}
\usage{
snapshot_studcomp(
  snapshot_date,
  dm,
  studcomp = clindata::rawplus_studcomp,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{studcomp}{\code{data.frame} Study completion data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot STUDCOMP Domain
}
