% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_study}
\alias{snapshot_study}
\title{Snapshot CTMS Study Domain}
\usage{
snapshot_study(
  snapshot_date,
  dm = clindata::rawplus_dm,
  study = clindata::ctms_study,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{dm}{\code{data.frame} Demographic data.}

\item{study}{\code{data.frame} CTMS study data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot CTMS Study Domain
}
