% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_visdt}
\alias{snapshot_visdt}
\title{Snapshot VISDT Domain}
\usage{
snapshot_visdt(
  snapshot_date,
  visdt = clindata::rawplus_visdt,
  print_check_rows = TRUE
)
}
\arguments{
\item{snapshot_date}{\code{Date} Date at which to snapshot data}

\item{visdt}{\code{data.frame} Visit date data.}

\item{print_check_rows}{\code{logical} Print row-checking output to console? Default: \code{TRUE}.}
}
\description{
Snapshot VISDT Domain
}
