% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cache.R
\name{cache_init}
\alias{cache_init}
\alias{cache_ls}
\alias{cache_dir}
\alias{cache_destroy}
\title{Data Caching API}
\usage{
cache_init(path = cache_dir())

cache_ls(path = cache_dir(), ...)

cache_dir()

cache_destroy(path = cache_dir(), verbose = TRUE)
}
\arguments{
\item{path}{character, path to cache directory, Default: \code{cache_dir()}}

\item{\dots}{arguments passed to \code{\link[=list.files]{list.files()}}}

\item{verbose}{logical, show message on success of \code{\link[=cache_destroy]{cache_destroy()}}}
}
\value{
character
}
\description{
Functions to manage cache data across sessions.
}
\details{
If the environment variable CLINDATA_CACHE_DIR is set then \code{\link[=cache_dir]{cache_dir()}} will use the value.

If the environment variable CLINDATA_CACHE_DIR is not set then \code{\link[=cache_dir]{cache_dir()}} will
default to the internal mechanism via \link[tools:userdir]{R_user_dir}.
}
\examples{
 cache_dir()
 cache_init()
 cache_ls()
 cache_destroy()
}
