% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms_2_process.R, R/make_ctms.R
\name{ctms_2_process}
\alias{ctms_2_process}
\title{Process CTMS data}
\usage{
ctms_2_process(datasets)

ctms_2_process(datasets)
}
\arguments{
\item{datasets}{\code{list} named list of data frames}
}
\description{
Process CTMS data

Process CTMS data
}
\keyword{internal}
