% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctms.R
\docType{data}
\name{ctms_study}
\alias{ctms_study}
\title{Study Metadata}
\format{
a data frame with 1 rows and 22 columns
}
\source{
./data-raw/ctms/study.csv
}
\usage{
ctms_study
}
\description{
Study Metadata (study) data CTMS data
}
\keyword{datasets}
