% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc.R
\docType{data}
\name{edc_data_entry_lag}
\alias{edc_data_entry_lag}
\title{Data Entry Lag}
\format{
a data frame with 235499 rows and 5 columns
}
\source{
./data-raw/edc/data_entry_lag.csv
}
\usage{
edc_data_entry_lag
}
\description{
data page-level data with simulated data entry lag (data_entry_lag)
}
\keyword{datasets}
