% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc.R
\docType{data}
\name{edc_queries}
\alias{edc_queries}
\title{Queries}
\format{
a data frame with 42259 rows and 13 columns
}
\source{
./data-raw/edc/queries.csv
}
\usage{
edc_queries
}
\description{
query-level data with simulated query characteristics (queries)
}
\keyword{datasets}
