% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRawplus.R
\name{map_rawplus_pd}
\alias{map_rawplus_pd}
\title{map_rawplus_pd}
\usage{
map_rawplus_pd(dfRawPD, ids)
}
\arguments{
\item{dfRawPD}{data frame of raw AE data}
}
\description{
mapping to create rawplus_pd from raw
}
\examples{

rawplus_pd <- map_rawplus_pd(clindata::raw_protdev)
}
