% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rawplus.R
\name{rawplus_2_process}
\alias{rawplus_2_process}
\title{Process \link{ rawplus } data}
\usage{
rawplus_2_process(datasets)
}
\arguments{
\item{datasets}{\code{list} named list of data frames}
}
\description{
Process \link{ rawplus } data
}
\keyword{internal}
