% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_consent}
\alias{rawplus_consent}
\title{rawplus_consent}
\format{
a data frame with 1297 rows and 4 columns.
}
\usage{
rawplus_consent
}
\description{
Informed consent data in Raw+ format
}
\keyword{datasets}
