% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_dm}
\alias{rawplus_dm}
\title{rawplus_dm}
\format{
a data frame with 1301 rows and 47 columns.
\describe{
\item{studyid}{\link{ string } Study Identifier}
\item{siteid}{\link{ string } Site Identifier}
\item{invid}{\link{ string } Investigator ID}
\item{scrnid}{\link{ string } Screen Identifier}
\item{subjid}{\link{ string } Subject Identifier}
\item{subjectid}{\link{ string } System Subject Identifier}
\item{datapagename}{\link{ string } eCRF Page Name}
\item{datapageid}{\link{ string } Internal ID for Data Page}
\item{foldername}{\link{ string } Folder Name}
\item{instancename}{\link{ string } Instance Name}
\item{recordid}{\link{ string } Record Identifier}
\item{record_dt}{\link{ string } Record Date}
\item{recordposition}{\link{ float } Record Position}
\item{mincreated_dts}{\link{ string } Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\link{ string } Latest Data Update Date/Time}
\item{brthdat}{\link{ string } Date of Birth}
\item{agerep}{\link{ string } Reported Age}
\item{sex}{\link{ string } Sex}
\item{ethnic}{\link{ string } Ethnicity}
\item{race}{\link{ string } Race}
\item{raceoth}{\link{ string } Race Other}
\item{racescat}{\link{ string } Race Sub-Category}
\item{racesoth}{\link{ string } Race Sub-Category Other}
\item{subjid_nsv}{\link{ string } SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\link{ string } SCREENING IDENTIFIER}
\item{subjinit_nsv}{\link{ string } SUBJECT INITIALS}
\item{invid_nsv}{\link{ string } INVESTIGATOR ID}
\item{subject_nsv}{\link{ string } Subject name or identifier}
\item{instanceid_nsv}{\link{ double } Internal id for the instance}
\item{folder_nsv}{\link{ string } Folder OID}
\item{folderseq_nsv}{\link{ double } Folder sequence number}
\item{brthdat_nsv}{\link{ timestamp } Date of Birth}
\item{brthdat_yy_nsv}{\link{ double } Date of Birth_Year}
\item{brthdat_mm_nsv}{\link{ double } Date of Birth_Month}
\item{brthdat_dd_nsv}{\link{ double } Date of Birth_Day}
\item{age_nsv}{\link{ double } Age (in years)}
\item{sex_std_nsv}{\link{ string } Sex at birth_Coded Value}
\item{ethnic_std_nsv}{\link{ string } Ethnicity_Coded Value}
\item{race_std_nsv}{\link{ string } Race_Coded Value}
\item{racescat_std_nsv}{\link{ string } Race Sub-category_Coded Value}
\item{firstdosedate}{\link{ character }}
\item{lastdosedate}{\link{ character }}
\item{lastparticipantdate}{\link{ character }}
\item{firstparticipantdate}{\link{ character }}
\item{timeontreatment}{\link{ int } Time on Treatment}
\item{timeonstudy}{\link{ int } Time on Study}
\item{country}{\link{ character }}
}
}
\source{
./data-raw/rawplus/dm.parquet
}
\usage{
rawplus_dm
}
\description{
Demographics (dm) data in \link{ rawplus } format.
}
\keyword{datasets}
