% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_ie}
\alias{rawplus_ie}
\title{rawplus_ie}
\format{
a data frame with 30168 rows and 4 columns.
}
\source{
./data-raw/rawplus/rawplus_ie.rda
}
\usage{
rawplus_ie
}
\description{
Visits (visdt) data in \code{rawplus} format.
\describe{
\item{subjid}{\code{string} Study subject identifier}
\item{iecat}{\code{string} Indicator for inclusion (INCL) or exclusion (EXCL)}
\item{ieorres}{\code{character} Indicator for inclusion (1) or exclusion (0)}
\item{tiver}{\code{string} Trial version indicator}
}
}
\keyword{datasets}
