% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_ixrsrand}
\alias{rawplus_ixrsrand}
\title{rawplus_ixrsrand}
\format{
a data frame with 1301 rows and 18 columns.
\describe{
\item{studyid}{\link{ string }}
\item{invid}{\link{ string }}
\item{invnam}{\link{ string }}
\item{oinvid}{\link{ string }}
\item{oinvnam}{\link{ string }}
\item{country}{\link{ string }}
\item{ocountry}{\link{ string }}
\item{txtyp_nsv}{\link{ string }}
\item{scrnid_nsv}{\link{ string }}
\item{subjid}{\link{ string }}
\item{subjinit_nsv}{\link{ string }}
\item{hbvdna_nsv}{\link{ string }}
\item{virtreat_nsv}{\link{ string }}
\item{trtgrpc}{\link{ string }}
\item{randnum}{\link{ string }}
\item{dispid_nsv}{\link{ double }}
\item{status}{\link{ string }}
\item{rand_dt}{\link{ string }}
}
}
\source{
./data-raw/rawplus/ixrsrand.parquet
}
\usage{
rawplus_ixrsrand
}
\description{
Randomization (ixrsrand) data in \link{ rawplus } format.
}
\keyword{datasets}
