% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_lb}
\alias{rawplus_lb}
\title{rawplus_lb}
\format{
a data frame with 1132070 rows and 13 columns.
}
\source{
./data-raw/rawplus/lb.parquet
}
\usage{
rawplus_lb
}
\description{
Labs (lb) data in \code{rawplus} format.
\describe{
\item{subjid}{\code{string} Subject ID or Number}
\item{visnam}{\code{string} Visit Name}
\item{visnum}{\code{float} Visit ID or Number}
\item{lb_dt}{\code{string} Actual Collection Date and Time}
\item{battrnam}{\code{string} Battery Name}
\item{lbtstnam}{\code{string} Test Name ID}
\item{siresn}{\code{float} SI Numeric Result}
\item{sinrlo}{\code{string} SI Reference Range Low}
\item{sinrhi}{\code{string} SI Reference Range High}
\item{toxgrg_nsv}{\code{string} Toxicity Grade}
\item{alrtfl}{\code{string} Alert Flag}
\item{treatmentemergent}{\code{character}}
\item{alertsimplified}{\code{character}}
}
}
\keyword{datasets}
