% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_lb}
\alias{rawplus_lb}
\title{rawplus_lb}
\format{
a data frame with 2332541 rows and 76 columns.
\describe{
\item{studyid}{\link{ string } Study ID or Number}
\item{invid}{\link{ string } Investigator ID or Number}
\item{scrnid}{\link{ string } Screen ID or Number}
\item{subjid}{\link{ string } Subject ID or Number}
\item{filcrdtm}{\link{ string } File Creation Date and Time}
\item{trmtyp}{\link{ string } Transmission Type}
\item{invnam}{\link{ string } Investigator Name}
\item{sex}{\link{ string } Subject Sex}
\item{brth_dt}{\link{ string } Subject Date Of Birth}
\item{visnum}{\link{ float } Visit ID or Number}
\item{visnam}{\link{ string } Visit Name}
\item{visittyp}{\link{ string } Visit Type}
\item{visitmod}{\link{ string } Visit Type Modifier}
\item{lbnam}{\link{ string } Central Laboratory Name}
\item{accsnnum}{\link{ string } Accession ID or Number}
\item{lactdtm}{\link{ string } Last Active Date and Time}
\item{recextyp}{\link{ string } Record Extension Type}
\item{lb_dt}{\link{ string } Actual Collection Date and Time}
\item{ptmel}{\link{ string } Planned Collection Time Elapsed}
\item{ptmeltx}{\link{ string } Planned Collection Time Elapsed Desc}
\item{colendtm}{\link{ string } Collection End Date and Time}
\item{rcvdtm}{\link{ string } Received Date and Time}
\item{speccnd}{\link{ string } Specimen Condition}
\item{speccom}{\link{ string } Lab - Specimen Comments}
\item{specicom}{\link{ string } Investigator - Specimen Comments}
\item{specmnum}{\link{ string } Specimen Material ID}
\item{specnam}{\link{ string } Specimen Material Name}
\item{ageatcol}{\link{ float } Subject Age at Collection}
\item{ageu}{\link{ string } Subject Age Units}
\item{faststat}{\link{ string } Fasting Status}
\item{battrid}{\link{ string } Battery ID}
\item{battrnam}{\link{ string } Battery Name}
\item{lbtstcd}{\link{ string } Test ID}
\item{lbtstnam}{\link{ string } Test Name ID}
\item{lbloinc}{\link{ string } LOINC Code}
\item{loinccd}{\link{ string } LOINC Code List ID}
\item{tststat}{\link{ string } Test Status}
\item{tstcom}{\link{ string } Test Level Comments}
\item{tsttyp}{\link{ string } Test Type}
\item{rptresc}{\link{ string } Reported Text Result}
\item{rptresn}{\link{ float } Reported Numeric Result}
\item{rptnrlo}{\link{ string } Reported Reference Range Low}
\item{rptnrhi}{\link{ string } Reported Reference Range High}
\item{rptu}{\link{ string } Reported Units}
\item{cnvresc}{\link{ string } Conventional Text Result}
\item{cnvresn}{\link{ float } Conventional Numeric Result}
\item{cnvnrhi}{\link{ string } Conventional Reference Range High}
\item{cnvu}{\link{ string } Conventional Units}
\item{siresc}{\link{ string } SI Text Result}
\item{siresn}{\link{ float } SI Numeric Result}
\item{sinrlo}{\link{ string } SI Reference Range Low}
\item{sinrhi}{\link{ string } SI Reference Range High}
\item{siu}{\link{ string } SI Units}
\item{rptrtyp}{\link{ string } Reported Result Type}
\item{rptrstat}{\link{ string } Reported Result Status}
\item{alrtfl}{\link{ string } Alert Flag}
\item{deltfl}{\link{ string } Delta Flag}
\item{toxgr}{\link{ string } Toxicity Grade}
\item{exclfl}{\link{ string } Exclusion Flag}
\item{blndfl}{\link{ string } Blinding Flag}
\item{rptdtm}{\link{ string } Reported Date and Time}
\item{trnstyp}{\link{ string } Transaction Type}
\item{subjinit_nsv}{\link{ string } SUBJECT INITIALS}
\item{brthdtn_nsv}{\link{ string } SUBJECT DATE OF BIRTH NUMERIC}
\item{lbdtn_nsv}{\link{ string } ACTUAL COLLECTION DATE NUMERIC}
\item{cnvnrlo_nsv}{\link{ string } CONVENTIONAL REFERENCE RANGE LOW}
\item{specmcd_nsv}{\link{ string } SPECIMEN MATERIAL CODE LIST ID}
\item{toxgrdsc_nsv}{\link{ string } Toxicity Grade Description}
\item{toxgrver_nsv}{\link{ string } Toxicity Grade Scale Version}
\item{toxgrg_nsv}{\link{ string } Toxicity Grade}
\item{toxgrnam_nsv}{\link{ string } Toxicity Grade Scale Name}
\item{toxgrhyp_nsv}{\link{ string } Toxicity Grade Direction}
}
}
\source{
C:/Users/schildress/dev/clindata/data-raw/rawplus/lb.parquet
}
\usage{
rawplus_lb
}
\description{
Labs (lb) data in Raw+ format.
}
\keyword{datasets}
