% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_sdrgcomp}
\alias{rawplus_sdrgcomp}
\title{rawplus_sdrgcomp}
\format{
a data frame with 689 rows and 34 columns.
\describe{
\item{studyid}{\link{ string } Study Identifier}
\item{siteid}{\link{ string } Site Identifier}
\item{invid}{\link{ string } Investigator ID}
\item{scrnid}{\link{ string } Screen Identifier}
\item{subjid}{\link{ string } Subject Identifier}
\item{subjectid}{\link{ string } System Subject Identifier}
\item{datapagename}{\link{ string } eCRF Page Name}
\item{datapageid}{\link{ string } Internal ID for Data Page}
\item{foldername}{\link{ string } Folder Name}
\item{instancename}{\link{ string } Instance Name}
\item{recordid}{\link{ string } Record Identifier}
\item{record_dt}{\link{ string } Record Date}
\item{recordposition}{\link{ float } Record Position}
\item{mincreated_dts}{\link{ string } Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\link{ string } Latest Data Update Date/Time}
\item{sdrgyn}{\link{ string } Did subject complete study drug dosing?}
\item{sdrgreas}{\link{ string } Reason for study drug discontinue}
\item{sdrgterm}{\link{ string } Discontinuation reasons}
\item{subjid_nsv}{\link{ string } SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\link{ string } SCREENING IDENTIFIER}
\item{subjinit_nsv}{\link{ string } SUBJECT INITIALS}
\item{invid_nsv}{\link{ string } INVESTIGATOR ID}
\item{subject_nsv}{\link{ string } Subject name or identifier}
\item{instanceid_nsv}{\link{ double } Internal id for the instance}
\item{folder_nsv}{\link{ string } Folder OID}
\item{folderseq_nsv}{\link{ double } Folder sequence number}
\item{sdrgyn_std_nsv}{\link{ string } Open label study drug completed_Coded Value}
\item{sdrgreas_std_nsv}{\link{ string } Reason for study drug discontinuation_Coded Value}
\item{sdrgol_nsv}{\link{ string } Will subject go to Open-Label}
\item{sdrgol_std_nsv}{\link{ string } Will subject go to Open-Label_Coded Value}
\item{sdrgfu_nsv}{\link{ string } Will subject enter 24-week trt free fup}
\item{sdrgfu_std_nsv}{\link{ string } Will subject enter 24-week trt free fup_Coded Value}
\item{sdrgtrt_nsv}{\link{ string } Did subject start another HBV therapy}
\item{sdrgtrt_std_nsv}{\link{ string } Did subject start another HBV therapy_Coded Value}
}
}
\source{
./data-raw/rawplus/sdrgcomp.parquet)}
}
\usage{
rawplus_sdrgcomp
}
\description{
Treatment Disposition (sdrgcomp) data in Raw+ format.
}
\keyword{datasets}
