% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplus.R
\docType{data}
\name{rawplus_studcomp}
\alias{rawplus_studcomp}
\title{rawplus_studcomp}
\format{
a data frame with 133 rows and 31 columns.
\describe{
\item{studyid}{\link{ string } Study Identifier}
\item{siteid}{\link{ string } Site Identifier}
\item{invid}{\link{ string } Investigator ID}
\item{scrnid}{\link{ string } Screen Identifier}
\item{subjid}{\link{ string } Subject Identifier}
\item{subjectid}{\link{ string } System Subject Identifier}
\item{datapagename}{\link{ string } eCRF Page Name}
\item{datapageid}{\link{ string } Internal ID for Data Page}
\item{foldername}{\link{ string } Folder Name}
\item{instancename}{\link{ string } Instance Name}
\item{recordid}{\link{ string } Record Identifier}
\item{record_dt}{\link{ string } Record Date}
\item{recordposition}{\link{ float } Record Position}
\item{mincreated_dts}{\link{ string } Earliest Data Creation Date/Time}
\item{maxupdated_dts}{\link{ string } Latest Data Update Date/Time}
\item{compyn}{\link{ string } Subject complete duration of the study?}
\item{compreas}{\link{ string } Reason for study discontinue}
\item{subjid_nsv}{\link{ string } SUBJECT IDENTIFIER}
\item{scrnid_nsv}{\link{ string } SCREENING IDENTIFIER}
\item{subjinit_nsv}{\link{ string } SUBJECT INITIALS}
\item{invid_nsv}{\link{ string } INVESTIGATOR ID}
\item{subject_nsv}{\link{ string } Subject name or identifier}
\item{instanceid_nsv}{\link{ double } Internal id for the instance}
\item{folder_nsv}{\link{ string } Folder OID}
\item{folderseq_nsv}{\link{ double } Folder sequence number}
\item{compyn_std_nsv}{\link{ string } Subject complete duration of the study?_Coded Value}
\item{compreas_std_nsv}{\link{ string } If No, reason for study discontinuation_Coded Value}
\item{compfu_nsv}{\link{ string } 24-week trt free fup completed}
\item{compfu_std_nsv}{\link{ string } 24-week trt free fup completed_Coded Value}
\item{comptrt_nsv}{\link{ string } Did subject start another HBV therapy}
\item{comptrt_std_nsv}{\link{ string } Did subject start another HBV therapy_Coded Value}
}
}
\source{
./data-raw/rawplus/studcomp.parquet)}
}
\usage{
rawplus_studcomp
}
\description{
Study Disposition (studcomp) data in Raw+ format.
}
\keyword{datasets}
